Move observability to pkg
